refactoring the code to isolate the converters methods into a package

+++ src/main/java/br/com/bruno/ecommerce/controllers/ClientController.java
@@ -35,7 +35,7 @@ public class ClientController {
     }
 
     @DeleteMapping
-    public ResponseEntity delete(@PathVariable("clientId") Integer clientId) {
+    public ResponseEntity delete(@PathVariable("clientId") Long clientId) {
 
         clientService.delete(clientId);
